Create driver.h
This is the header file which defines class Driver, which reads the input file, calls the Work module to perform the transformation, then displays the output.  This handles the various command modes, like the unit test mode which reads an input line, calls Work to process it, reads the next line and compares it to the output string, reporting an error if they don't match, then loops until the end of the file.